+++ source/test-suite/fixtures/TypeOperatorRecordInvalid.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE TypeOperators #-}
+module TypeOperatorRecordInvalid where
+
+data Broken = Broken
+  { field :: left :::
+  }